All: port python3
git-svn-id: https://elsa-svn.onera.fr/CASSIOPEE/Trunk/Cassiopee/Apps/Modules@3246 2bbd7fe0-4929-0410-bb6b-e55522daae51

+++ KCore/KCore/Fld/ngon_t.hxx
@@ -258,7 +258,7 @@ struct ngon_t
       // sort the nodes
       BARSplitter::getSortedNodes(cB, snodes);
       
-      if (snodes.size() != nb_edges) //degen
+      if (snodes.size() != (size_t)nb_edges) //degen
       {
         //std::cout << cB << std::endl;
         continue;
